+++ src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.appid.spring.boot.AppIDOAuth2AutoConfig
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.ibm.boot.autoconfigure.appid.AppIDOAuth2AutoConfig
